Add files which the other two endpoints need
The popularity technique endpoint needs an indices file and a
cosine similarity matrix file. That added to the repo a while ago.
This commit adds similar files which are needed in the other two
endpoints. However, I could not run this reliably on the iLab machines
because the process took up too much RAM. Even if I restarted the
server, it took up too much RAM, and it crashed. Also, it would
hang on the subprocess call to sed even though the sed call would
work correctly when manually called via the terminal.

I'll work on this further and fix these bugs.