Correct shell command that runs popularity and keyword endpoints
The shell command was wrong, and it was very confusing. I don't
know what it was trying to do. I replaced it, and now the endpoints
run in a second or two each. The endpoints give reasonable
recommendations. For example, if someone asks for a recommendation
based on GoldenEye, the endpoints recommend other James Bond movies.