Hide recommendations for partially-complete movie names
If you want to write 'Avatar', you have to write 'A' and 'Ava' on
your way there. There is a movie called 'A' and a movie called
'Ava', so the system would load recommendations for those movies
before loading them for 'Avatar'. However, if the user continues
to write something after 'A' or 'Ava', then they clearly don't
want to see those recommendations. This commit makes it so that,
when the recommendations are loaded from the server, the client
checks whether the query movie equals the movie currently in the
input field. If it is the same, it continues as normal. If they
are different, then the client doesn't display the movie recommendations
for that query.